Berry move mapping from C++ to C
