+++ lib/libesp32/berry_tasmota/src/be_display_lib.c
@@ -12,7 +12,7 @@
 
 extern int be_ntv_display_start(bvm *vm);
 extern int be_ntv_display_dimmer(bvm *vm);
-extern bool be_ntv_display_started(void);
+extern bbool be_ntv_display_started(void);
             BE_FUNC_CTYPE_DECLARE(be_ntv_display_started, "b", "")
 extern void be_ntv_display_touch_update(int32_t touches, int32_t raw_x, int32_t raw_y, int32_t gesture);
             BE_FUNC_CTYPE_DECLARE(be_ntv_display_touch_update, "", "iiii")
